@@ -5,73 +5,31 @@
 // To keep your imports tidy, follow the ordering guidelines at
 // https://www.dartlang.org/guides/language/effective-dart/style#ordering
 import 'package:flutter/material.dart';
-import 'package:flutter_app/ui/two/Unit.dart';
-// @required is defined in the meta.dart package
 import 'package:meta/meta.dart';
 
-import 'converter_route.dart';
-
+import 'MyCategory.dart';
 
 // We use an underscore to indicate that these variables are private.
 // See https://www.dartlang.org/guides/language/effective-dart/design#libraries
-final _rowHeight = 100.0;
+const _rowHeight = 100.0;
 final _borderRadius = BorderRadius.circular(_rowHeight / 2);
 
-/// A custom [Category] widget.
-///
-/// The widget is composed on an [Icon] and [Text]. Tapping on the widget shows
-/// a colored [InkWell] animation.
-class Category extends StatelessWidget {
-  final String name;
-  final ColorSwatch color;
-  final IconData iconLocation;
-  final List<Unit> units;
+/// A [CategoryTile] to display a [Category].
+class CategoryTile extends StatelessWidget {
+  final MyCategory category;
+  final ValueChanged<MyCategory> onTap;
 
-  /// Creates a [Category].
+  /// The [CategoryTile] shows the name and color of a [Category] for unit
+  /// conversions.
   ///
-  /// A [Category] saves the name of the Category (e.g. 'Length'), its color for
-  /// the UI, and the icon that represents it (e.g. a ruler).
-  // While the @required checks for whether a named parameter is passed in,
-  // it doesn't check whether the object passed in is null. We check that
-  // in the assert statement.
-  const Category({
+  /// Tapping on it brings you to the unit converter.
+  const CategoryTile({
     Key key,
-    @required this.name,
-    @required this.color,
-    @required this.iconLocation,
-    @required this.units,
-  })  : assert(name != null),
-        assert(color != null),
-        assert(iconLocation != null),
-        assert(units != null),
+    @required this.category,
+    @required this.onTap,
+  })  : assert(category != null),
+        assert(onTap != null),
         super(key: key);
-
-  /// Navigates to the [ConverterRoute].
-  void _navigateToConverter(BuildContext context) {
-    Navigator.of(context).push(MaterialPageRoute<Null>(
-      builder: (BuildContext context) {
-        return Scaffold(
-          appBar: AppBar(
-            elevation: 1.0,
-            title: Text(
-              name,
-              style: Theme.of(context).textTheme.display1,
-            ),
-            centerTitle: true,
-            backgroundColor: color,
-          ),
-          body: ConverterRoute(
-            color: color,
-            name: name,
-            units: units,
-          ),
-          // This prevents the attempt to resize the screen when the keyboard
-          // is opened
-          resizeToAvoidBottomPadding: false,
-        );
-      },
-    ));
-  }
 
   /// Builds a custom widget that shows [Category] information.
   ///
@@ -88,12 +46,11 @@ class Category extends StatelessWidget {
         height: _rowHeight,
         child: InkWell(
           borderRadius: _borderRadius,
-          // TODO: Use the highlight and splash colors from the ColorSwatch
-          highlightColor: color,
-          splashColor: color,
+          highlightColor: category.color['highlight'],
+          splashColor: category.color['splash'],
           // We can use either the () => function() or the () { function(); }
           // syntax.
-          onTap: () => _navigateToConverter(context),
+          onTap: () => onTap(category),
           child: Padding(
             padding: EdgeInsets.all(8.0),
             child: Row(
@@ -106,13 +63,13 @@ class Category extends StatelessWidget {
                 Padding(
                   padding: EdgeInsets.all(16.0),
                   child: Icon(
-                    iconLocation,
+                    category.iconLocation,
                     size: 60.0,
                   ),
                 ),
                 Center(
                   child: Text(
-                    name,
+                    category.name,
                     textAlign: TextAlign.center,
                     style: Theme.of(context).textTheme.headline,
                   ),
